Internal API

PRASInterface.CC_RESTRICTIONS_UTIL_FILE — Constant

FILE with restrictions on CC units in PRAS

PRASInterface.HVDCLineTypes — Type

Supported DC branch Types

PRASInterface.OUTAGE_INFO_FILE — Constant

DEFAULT outage data which is used when outage_flag is set to FALSE

Based on ERCOT historical data

PRASInterface.TransformerTypes — Constant

Filtered Transformer Types

These transformers are not modeled as lines in PRAS.

PRASInterface.S2P_metadata — Type

S2P_metadata

Struct to store metadata for the Sienna to PRAS conversion

PRASInterface.outage_data — Type

outage_data

Struct to store the outage information for a component.

PRASInterface.check_file — Method

check_file(loc::String)

Check if the file exists and is openable.

PRASInterface.get_available_components_in_aggregation_topology — Method

get_available_components_in_aggregation_topology(
    type::Type{<:PSY.StaticInjection},
    sys::PSY.System,
    region::PSY.AggregationTopology,
)

Get available components in the AggregationTopology region of the given type.

PRASInterface.get_first_ts — Method

get_first_ts(ts::TS)

Get time series and handle components with no time series (such as unavailable)

PRASInterface.get_generator_category — Method

get_generator_category(gen::StaticInjection)

Get the category of the generator.

Arguments

gen::StaticInjection: Generator

Returns

String: Category of the generator

PRASInterface.get_sorted_lines — Method

get_sorted_lines(lines::Vector{PSY.Branch}, region_names::Vector{String})

Get sorted lines, interface region indices, and interface line indices.

Arguments

lines::Vector{PSY.Branch}: Lines

region_names::Vector{String}: Region names

Returns

sorted_lines::Vector{PSY.Branch}: Sorted lines

interface_reg_idxs::Vector{Tuple{Int, Int}}: Interface region indices

interface_line_idxs::Vector{UnitRange{Int}}: Interface line indices

PRASInterface.get_sorted_region_tuples — Method

get_sorted_region_tuples(lines::Vector{PSY.Branch}, region_names::Vector{String})

Get sorted (regfrom, regto) tuples of inter-regional lines.

PRASInterface.get_timestamps — Method

get_timestamps(initial_time, resolution, steps, horizon)

Return a vector of timestamps starting from initial_time with resolution and steps for horizon steps

PRASInterface.get_ts_values — Method

get_ts_values(ts::PSY.TimeSeriesData)

Get the time series values.

Arguments

ts::TimeSeries: Time series

Returns

Array{Float64}: Time series values

PRASInterface.initialize_availability! — Method

initializeavailability!(rng, availability, nexttransition, devices, tlast)

Arguments

rng::Random.AbstractRNG: Random number generator

availability::Vector{Bool}: Vector of availability status

nexttransition::Vector{Int}: Vector of next transition time

devices::Vector{PSY.Generator}: Vector of devices with outage SupplementalAttributes

t_last::Int: Last time step

Returns

Modifies input arguments and rereturns

availability::Vector{Bool}: Vector of availability status

nexttransition::Vector{Int}: Vector of next transition time

PRASInterface.line_rating — Method

line_rating(line::Branch)

Get the line rating.

Arguments

line::Branch: Line

Returns

Tuple{forward_capacity::Float64, backward_capacity::Float64}: Line rating

PRASInterface.make_pras_interfaces — Method

make_pras_interfaces(
    sorted_lines::Vector{PSY.Branch},
    interface_reg_idxs::Vector{Tuple{Int64, Int64}},
    interface_line_idxs::Vector{UnitRange{Int64}},
    s2p_meta::S2P_metadata,
)

Converts PSY branches and interaces indices into PRAS Lines and Interfaces.

Returns

new_lines::PRAS.Lines: PRAS Lines

new_interfaces::PRAS.Interfaces: PRAS Interfaces

PRASInterface.rate_to_probability — Method

rate_to_probability(for_gen::Float64, mttr::Int64)

Converts the forced outage rate and mean time to repair to the λ and μ parameters

Arguments

for_gen::Float64: Forced outage rate [1/T]

mttr::Int64: Mean time to repair [T]

Returns

λ::Float64: Transition probability from online to offline [1/T]

μ::Float64: Transition rate from offline to online [1/T]

Reference

https://core.ac.uk/download/pdf/13643059.pdf

PRASInterface.runchecks — Method

runchecks(sys_location::String)

Check if the System JSON file is serialized as well as other files required.

PRASInterface.update_availability! — Method

update_availability!(rng, availability, nexttransition, devices, t_now, t_last)

Return availability and next transition with new randomness