Public API Reference

PRASInterface.generate_csv_outage_profile — Method

generateoutageprofile(prassystem,numruns,psysys,numscenarios,location)

Process the assess results to get timeseries of generator status and include this timeseries data to the corresponding component in PSY System and exported using to_json method (serializing the PSY System).

...

Arguments

pras_system::PRAS.SystemModel: PRAS System

num_runs::Int64: Number of PRAS runs

psy_sys::PSY.System: PSY System

num_scenarios::Int64: Number of scenarios of user interest.

location::String: Location to store outage profile. ...

Examples

julia> generate_outage_profile(results, pras_sys, psy_sys, 1)
PSY System exported using to_json method in InfrastructureSystems

PRASInterface.generate_outage_profile — Method

generateoutageprofile(prassystem,numruns,psysys,numscenarios,location)

Process the assess results to get timeseries of generator status and include this timeseries data to the corresponding component in PSY System and exported using to_json method (serializing the PSY System).

...

Arguments

pras_system::PRAS.SystemModel: PRAS System

num_runs::Int64: Number of PRAS runs

psy_sys::PSY.System: PSY System

num_scenarios::Int64: Number of scenarios of user interest.

location::String: Location to store outage profile. ...

Examples

julia> generate_outage_profile(results, pras_sys, psy_sys, 1)
PSY System exported using to_json method in InfrastructureSystems

PRASInterface.generate_pras_system — Method

generate_pras_system(sys::PSY.System, aggregation; kwargs...)

Sienna/Data PowerSystems.jl System is the input and an object of PRAS SystemModel is returned. ...

Arguments

sys::PSY.System: Sienna/Data PowerSystems.jl System

aggregation<:PSY.AggregationTopology: "PSY.Area" (or) "PSY.LoadZone" {Optional}

availability::Bool: Takes into account avaialability of StaticInjection components when building the PRAS System {Optional}

lump_region_renewable_gens::Bool: Whether to lumps PV and Wind generators in a region because usually these generators don't have FOR data {Optional}

export_location::String: Export location of the .pras file ...

Returns

- `PRAS.SystemModel`: PRAS SystemModel object

Examples

julia> make_pras_system(psy_sys)
PRAS SystemModel