Public API Reference

PRASInterface — Module

PowerSystems Interface for Probabilistic Resource Adequacy Studies (PRAS)

Key Functions

generate_pras_system: convert PSY to PRAS model

assess: assess PRAS model

Key PRAS Types

SystemModel: PRAS data structure

SequentialMonteCarlo: method for PRAS analysis

Shortfall: PRAS metric for missing generation

LOLE: PRAS metric for loss of load expectation

EUE: PRAS metric for energy unserved expectation

PRASInterface.generate_pras_system — Function

generate_pras_system(sys::PSY.System, aggregation; kwargs...)

Sienna/Data PowerSystems.jl System is the input and an object of PRAS SystemModel is returned. ...

Arguments

sys::PSY.System: Sienna/Data PowerSystems.jl System

aggregation<:PSY.AggregationTopology: "PSY.Area" (or) "PSY.LoadZone" {Optional}

availability::Bool: Takes into account avaialability of StaticInjection components when building the PRAS System {Optional}

lump_region_renewable_gens::Bool: Whether to lumps PV and Wind generators in a region because usually these generators don't have FOR data {Optional}

export_location::String: Export location of the .pras file ...

Returns

- `PRAS.SystemModel`: PRAS SystemModel object

Examples

julia> generate_pras_system(psy_sys)
PRAS SystemModel

generate_pras_system(sys_location::String, aggregation; kwargs...)

Generate a PRAS SystemModel from a Sienna/Data PowerSystems System JSON file.

Arguments

sys_location::String: Location of the Sienna/Data PowerSystems System JSON file

aggregation::Type{AT}: Aggregation topology type

availability::Bool: Availability of components in the System

lump_region_renewable_gens::Bool: Lumping of region renewable generators

export_location::Union{Nothing, String}: Export location of the .pras file

Returns

PRAS.SystemModel: PRAS SystemModel

PRASInterface.PRAS — Module

PRAS

Module for Probabilistic Resource Adequacy Studies (PRAS).

Re-exported in PRASInterface

Source

https://github.com/NREL/PRAS.jl

PRASInterface.PRAS.PRASBase.SystemModel — Type

SystemModel

SystemModel is the primary data structure for Probabilistic Resource Adequacy Studies (PRAS).

You can also load a SystemModel from an appropriately-formatted HDF5 file on disk.

Examples

pras = SystemModel("path/to/pras.pras")

PRASInterface.PRAS.ResourceAdequacy.assess — Function

assess(system::SystemModel, method::SequentialMonteCarlo, resultspecs::ResultSpec...)

Run a Sequential Monte Carlo simulation on a system using the method data and return resultspecs.

Arguments

system::SystemModel: PRAS data structure

method::SequentialMonteCarlo: method for PRAS analysis

resultspecs::ResultSpec...: PRAS metric for metrics like Shortfall missing generation

Returns

results::Tuple{Vararg{ResultAccumulator{SequentialMonteCarlo}}}: PRAS metric results

PRASInterface.PRAS.ResourceAdequacy.SequentialMonteCarlo — Type

SequentialMonteCarlo(;
    samples::Int=10_000,
    seed::Integer=rand(UInt64),
    verbose::Bool=false,
    threaded::Bool=true
)

Sequential Monte Carlo simulation parameters for PRAS analysis

It it recommended that you fix the random seed for reproducibility.

Arguments

samples::Int=10_000: Number of samples

seed::Integer=rand(UInt64): Random seed

verbose::Bool=false: Print progress

threaded::Bool=true: Use multi-threading

Returns

SequentialMonteCarlo: PRAS analysis method

PRASInterface.PRAS.ResourceAdequacy.Shortfall — Type

Shortfall

Shortfall metric represents lost load at regions and timesteps in ShortfallResult with a (regions, timestamps) matrix API.

Separate samples are averaged together into mean and std values.

See ShortfallSamples for all shortfall samples.

PRASInterface.PRAS.ResourceAdequacy.Surplus — Type

Surplus

Surplus metric represents extra generation at regions and timestamps in a SurplusResults with a (regions, timestamps) matrix API.

Separate samples are averaged together into mean and std values.

See SurplusSamples for all surplus samples.

PRASInterface.PRAS.ResourceAdequacy.Flow — Type

Flow

Flow metric represents the flow between interfaces at timestamps in a FlowResult with a (interfaces, timestamps) matrix API.

Separate samples are averaged together into mean and std values.

See FlowSamples for all flow samples.

PRASInterface.PRAS.ResourceAdequacy.Utilization — Type

Utilization

Utilization metric represents how much an interface between regions is used across timestamps in a UtilizationResult with a (interfaces, timestamps) matrix API.

Separate samples are averaged together into mean and std values.

See UtilizationSamples for all utilization samples.

PRASInterface.PRAS.ResourceAdequacy.StorageEnergy — Type

StorageEnergy

Storage energy represents the state-of-charge of storage resources at timestamps in a StorageEnergyResult with a (storages, timestamps) matrix API.

Separate samples are averaged together into mean and std values.

See StorageEnergySamples for all storage energy samples.

See GeneratorStorageEnergy for generator storage energy.

PRASInterface.PRAS.ResourceAdequacy.GeneratorStorageEnergy — Type

GeneratorStorageEnergy

Generator storage energy represents state-of-charge of generatorstorage resources at timestamps in a StorageEnergyResult with a (generatorstorages, timestamps) matrix API.

Separate samples are averaged together into mean and std values.

See GeneratorStorageEnergySamples for all generator storage energy samples.

See StorageEnergy for storage energy.

PRASInterface.PRAS.ResourceAdequacy.LOLE — Type

LOLE

Loss of load expectation metric. Contains a mean and standard error estimate.

PRASInterface.PRAS.ResourceAdequacy.EUE — Type

EUE

Expected unserved energy expectation metric. Contains a mean and standard error estimate.

Missing docstring.

Missing docstring for val. Check Documenter's build log for details.

Missing docstring.

Missing docstring for stderror. Check Documenter's build log for details.

PRASInterface.generate_outage_profile — Function

generate_outage_profile(pras_system,num_runs,psy_sys,num_scenarios,location)

Process the assess results to get timeseries of generator status and include this timeseries data to the corresponding component in PSY System and exported using to_json method (serializing the PSY System).

...

Arguments

pras_system::PRAS.SystemModel: PRAS System

num_runs::Int64: Number of PRAS runs

psy_sys::PSY.System: PSY System

num_scenarios::Int64: Number of scenarios of user interest.

location::String: Location to store outage profile. ...

Examples

julia> generate_outage_profile(results, pras_sys, psy_sys, 1)
PSY System exported using to_json method in InfrastructureSystems

PRASInterface.generate_csv_outage_profile — Function

generate_outage_profile(pras_system,num_runs,psy_sys,num_scenarios,location)

Process the assess results to get timeseries of generator status and include this timeseries data to the corresponding component in PSY System and exported using to_json method (serializing the PSY System).

...

Arguments

pras_system::PRAS.SystemModel: PRAS System

num_runs::Int64: Number of PRAS runs

psy_sys::PSY.System: PSY System

num_scenarios::Int64: Number of scenarios of user interest.

location::String: Location to store outage profile. ...

Examples

julia> generate_outage_profile(results, pras_sys, psy_sys, 1)
PSY System exported using to_json method in InfrastructureSystems

PRASInterface.add_csv_time_series! — Function

add_csv_time_series!(
    sys_DA,
    sys_RT,
    outage_csv_location::String;
    days_of_interest::Union{Nothing, UnitRange}=nothing,
    add_scenario::Union{Nothing, Int}=nothing,
)

Generates outage profile for two stage PowerSimulation and adds availability time series data to Generators in PSY System from CSV files.

PRASInterface.add_csv_time_series_single_stage! — Function

add_csv_time_series_single_stage!(
    sys_DA,
    outage_csv_location::String;
    days_of_interest::Union{Nothing, UnitRange}=nothing,
    add_scenario::Union{Nothing, Int}=nothing,
)

Generates outage profile for single stage PowerSimulation and adds availability time series.

PRASInterface.make_generator_outage_draws! — Function

make_generator_outage_draws!(
    sys,
    initial_time::Dates.DateTime=nothing,
    resolution::TIMEPERIOD=nothing,
    steps::Int=nothing,
    horizon::Int=nothing,
) where {TIMEPERIOD <: Dates.TimePeriod}

Adds availability time series to the generators in the system.

Main function to make generator outage draws.

PRASInterface.PRAS.ResourceAdequacy.ShortfallSamples — Type

ShortfallSamples

ShortfallSamples metric represents lost load at regions and timesteps in ShortfallSamplesResult with a (regions, timestamps, samples) matrix API.

See Shortfall for averaged shortfall samples.

PRASInterface.PRAS.ResourceAdequacy.SurplusSamples — Type

SurplusSamples

Surplus samples represent extra generation at regions and timestamps in a SurplusSamplesResult with a (regions, timestamps, samples) matrix API.

See Surplus for sample-averaged surplus data.

PRASInterface.PRAS.ResourceAdequacy.FlowSamples — Type

FlowSamples

Flow samples represent the flow between interfaces at timestamps, which has not been averaged across different samples. This presents a 3D matrix API (interfaces, timestamps, samples).

See Flow for sample-averaged flow data.

PRASInterface.PRAS.ResourceAdequacy.UtilizationSamples — Type

UtilizationSamples

Utilization samples represent the utilization between interfaces at timestamps, which has not been averaged across different samples. This presents a 3D matrix API (interfaces, timestamps, samples).

See Utilization for averaged utilization samples.

PRASInterface.PRAS.ResourceAdequacy.StorageEnergySamples — Type

StorageEnergySamples

Storage energy samples represent the state-of-charge of storage resources at timestamps, which has not been averaged across different samples. This presents a 3D matrix API (storages, timestamps, samples).

See StorageEnergy for sample-averaged storage energy.

PRASInterface.PRAS.ResourceAdequacy.GeneratorStorageEnergySamples — Type

GeneratorStorageEnergySamples

Generator storage energy samples represent the state-of-charge of generatorstorage resources at timestamps, which has not been averaged across different samples. This presents a 3D matrix API (generatorstorages, timestamps, samples).

See GeneratorStorageEnergy for sample-averaged generator storage energy.

PRASInterface.PRAS.ResourceAdequacy.GeneratorAvailability — Type

GeneratorAvailability

Generator availability represents the availability of generators at timestamps in a GeneratorAvailabilityResult with a (generators, timestamps, samples) matrix API.

No averaging occurs.

PRASInterface.PRAS.ResourceAdequacy.GeneratorStorageAvailability — Type

GeneratorStorageAvailability

Generator storage availability represents the availability of generatorstorage resources at timestamps in a GeneratorStorageAvailabilityResult with a (generatorstorages, timestamps, samples) matrix API.

No averaging occurs

PRASInterface.PRAS.ResourceAdequacy.LineAvailability — Type

LineAvailability

Line availability represents the availability of lines at timestamps in a LineAvailabilityResult with a (lines, timestamps, samples) matrix API.

No averaging occurs.